requirements.txt
SQLAlchemy>=2.0.0
fastapi >= 0.89.1
uvicorn >= 0.20.0
PyMySQL~=1.0.2
pydantic~=1.10.4
python-multipart~=0.0.5
python-dotenv